src/test/extension.test.ts
import * as assert from 'assert';
import * as vscode from 'vscode';
// import * as myExtension from '../../extension';

suite('Extension Test Suite', () => {
    vscode.window.showInformationMessage('Start all tests.');

    setup(async () => {
        // Setup code before each test
        // Ensure a clean state if necessary, open files, etc.
        await vscode.commands.executeCommand('workbench.action.closeAllEditors');
    });

    test('Sample test', () => {
        assert.strictEqual(-1, [1, 2, 3].indexOf(5));
        assert.strictEqual(-1, [1, 2, 3].indexOf(0));
    });

    test('Extension Activation', async () => {
        const extension = vscode.extensions.getExtension('publisher.extensionName');
        assert.ok(extension);
        await extension.activate();
        assert.ok(extension.isActive);
    });

    test('Open Notebook File', async () => {
        const uri = vscode.Uri.file('/path/to/notebook.ipynb');
        const document = await vscode.workspace.openTextDocument(uri);
        const editor = await vscode.window.showTextDocument(document);
        assert.strictEqual(editor.document.languageId, 'jupyter');
    });

    test('Execute Shell Command in Notebook', async () => {
        const command = '!echo "Hello World"';
        const uri = vscode.Uri.file('/path/to/notebook.ipynb');
        await vscode.workspace.openTextDocument(uri).then(doc => {
            const editor = vscode.window.showTextDocument(doc);
            return editor.then(e => {
                e.edit(edit => {
                    edit.insert(new vscode.Position(0, 0), command);
                });
            });
        });

        // Simulate running the cell with the shell command
        await vscode.commands.executeCommand('nbsh.runShell');
        
        // Check output or effects of the command
        // This could involve checking the terminal output, files, or other side effects
        // As this is difficult to automate directly, consider logging or specific markers
    });

    test('Handle Invalid Shell Command', async () => {
        const command = '!invalidcommand';
        const uri = vscode.Uri.file('/path/to/notebook.ipynb');
        await vscode.workspace.openTextDocument(uri).then(doc => {
            const editor = vscode.window.showTextDocument(doc);
            return editor.then(e => {
                e.edit(edit => {
                    edit.insert(new vscode.Position(0, 0), command);
                });
            });
        });

        // Try to catch errors or handle expected failures
        try {
            await vscode.commands.executeCommand('nbsh.runShell');
            assert.fail("Expected an error for invalid command");
        } catch (error) {
            assert.ok("Caught expected error for invalid command");
        }
    });

    teardown(async () => {
        // Code to clean up after each test
        await vscode.commands.executeCommand('workbench.action.closeAllEditors');
    });
});